Middle update 06.03.2023 Templates, views and other changes.

+++ masters_app/views.py
@@ -1,9 +1,8 @@
 from django.contrib.auth import logout, login
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.views import LoginView
-from django.core.paginator import Paginator
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import HttpResponse
+from django.http import HttpResponse, Http404
 from django.urls import reverse_lazy
 from django.views import View
 from django.views.generic import DetailView, CreateView, ListView
@@ -52,7 +51,6 @@ class RegisterUser(DataMixin, CreateView):
     def get_context_data(self, *, object_list=None, **kwargs):
         context = super().get_context_data(**kwargs)
         context.update(self.get_user_context(title="Реєстрація"))
-        print(context)
         return context
 
     def form_valid(self, form):
@@ -83,28 +81,32 @@ class PageView(DataMixin, DetailView):
     def get_context_data(self, *, object_list=None, **kwargs):
         context = super().get_context_data(**kwargs)
         context['title'] = context['page']
-        context['albums'] = Album.objects.filter(page=context['page'])
-        context['links'] = Link.objects.filter(page=context['page'])
+        context['albums'] = Album.objects.filter(page=context['page'])[:3]
         context.update(self.get_user_context())
         return context
 
 
-class ChangePage(LoginRequiredMixin, DataMixin, View):
-    login_url = '/login/'
-    redirect_field_name = 'home'
+class ChangePage( DataMixin, View):
     template_name = 'masters_app/change_page.html'
 
     def dispatch(self, request,  *args, **kwargs):
-        page = get_object_or_404(Page, slug=kwargs['page_slug'])
+        user = request.user
+
+        # checking authentication
+        if not user.is_authenticated:
+            return redirect('login')
+
+        # POST
         if request.method == 'POST':
-            form = ChangePageForm(instance=request.user.page, data=request.POST, files=request.FILES)
+            form = ChangePageForm(instance=user.page, data=request.POST, files=request.FILES)
             if form.is_valid():
                 form.save()
                 return redirect('home')
 
+        # GET
         else:
-            form = ChangePageForm(instance=request.user.page)
-        context = self.get_user_context(form=form, title='Змінити сторінку', page=page)
+            form = ChangePageForm(instance=user.page)
+        context = self.get_user_context(form=form, title='Змінити сторінку', page=user.page)
         return render(request, self.template_name, context)
 
 
@@ -113,44 +115,59 @@ class SearchView(DataMixin, View):
     def get(self, request, *args, **kwargs):
         page_list = Page.objects.all()
         page_filter = MastersFilter(request.GET, queryset=page_list)
-        context = self.get_user_context(title='Beauty Masters')
-        context.update({'filter': page_filter})
+        searched = False
+        if '_search' in request.GET:
+            searched = True
+        context = self.get_user_context(title='Beauty Masters',
+                                        filter=page_filter,
+                                        searched=searched)
         return render(request, 'masters_app/search.html', context)
 
 
-class EditAccount(LoginRequiredMixin, DataMixin, View):
-    login_url = '/login/'
-    redirect_field_name = 'home'
+class EditAccount(DataMixin, View):
     template_name = 'masters_app/edit_account.html'
 
     def dispatch(self, request, *args, **kwargs):
-        page = get_object_or_404(Page, slug=kwargs['page_slug'])
+        user = request.user
+
+        # checking authentication
+        if not user.is_authenticated:
+            return redirect('login')
+
+        # POST
         if request.method == 'POST':
-            form = EditAccountForm(instance=request.user, data=request.POST)
+            form = EditAccountForm(instance=user, data=request.POST)
             if form.is_valid():
-                page.firstname = form.cleaned_data['first_name']
-                page.lastname = form.cleaned_data['last_name']
-                page.slug = form.cleaned_data['username'].lower()
-                page.save()
+                user.page.firstname = form.cleaned_data['first_name']
+                user.page.lastname = form.cleaned_data['last_name']
+                user.page.slug = form.cleaned_data['username'].lower()
+                user.page.save()
                 form.save()
                 return redirect('home')
+
+        # GET
         else:
-            form = EditAccountForm(instance=request.user)
-        context = self.get_user_context(form=form, title='Налаштування аккаунта', page=page)
+            form = EditAccountForm(instance=user)
+        context = self.get_user_context(form=form, title='Налаштування аккаунта', page=user.page)
         return render(request, self.template_name, context)
 
 
 class AlbumContentView(DataMixin, View):
     def get(self, request, *args, **kwargs):
 
-        page = get_object_or_404(Page, slug=kwargs['page_slug'])
         album = get_object_or_404(Album, pk=kwargs['album_id'])
+        page = album.page
+        owner = False
+        if request.user.page == page:
+            owner = True
         # paginator
         page_obj = self.get_paginator(request, Work, 8, album=album)
 
         context = self.get_user_context(title=f'{album.name}',
                                         page=page,
-                                        page_obj=page_obj
+                                        page_obj=page_obj,
+                                        album_id=album.pk,
+                                        owner=owner
                                         )
         return render(request, 'masters_app/album.html', context)
 
@@ -162,7 +179,6 @@ class WorkView(DataMixin, DetailView):
     context_object_name = 'work'
 
     def get_context_data(self, *, object_list=None, **kwargs):
-        print(kwargs)
         context = super().get_context_data(**kwargs)
         album = context['work'].album
         page = album.page
@@ -181,20 +197,23 @@ class WorkView(DataMixin, DetailView):
             prev_work = work.pk
 
         context.update(self.get_user_context(title=album,
-                                             queryset=queryset,
                                              page=page,
                                              prev_work=prev_work,
                                              next_work=next_work))
-
         return context
 
 
-class LinksView(LoginRequiredMixin, DataMixin, ListView):
+class LinksView(DataMixin, ListView):
     def dispatch(self, request, *args, **kwargs):
-        page = get_object_or_404(Page, slug=kwargs['page_slug'])
-        links = Link.objects.filter(page=page)
-        title = f'{page} links'
-        context = self.get_user_context(page=page,
+        user = request.user
+
+        # checking authentication
+        if not user.is_authenticated:
+            return redirect('login')
+
+        links = Link.objects.filter(page=user.page).select_related('link_type')
+        title = f'{user.page} links'
+        context = self.get_user_context(page=user.page,
                                         links=links,
                                         title=title)
         if request.method == 'POST':
@@ -211,46 +230,44 @@ class LinksView(LoginRequiredMixin, DataMixin, ListView):
                     form.save()
                 else:
                     print('Error')
-                return redirect('links', page_slug=page.slug)
+                return redirect('links')
 
             elif '_add_link' in request.POST:
                 form = LinkForm(data=request.POST)
                 if form.is_valid():
-                    Link.objects.create(page=page,
+                    Link.objects.create(page=user.page,
                                         link_type=form.cleaned_data['link_type'],
                                         url=form.cleaned_data['url'])
-                    return redirect('links', page_slug=page.slug)
+                    return redirect('links')
                 else:
-                    print(form.data)
-                    print(form.cleaned_data)
-                    print('Error')
+                    form.add_error(None, 'Unknown Error')
 
             elif '_add' in request.POST:
-                form = LinkForm(initial={'page': page})
+                form = LinkForm(initial={'page': user.page})
                 context.update(form=form)
 
             elif '_delete' in request.POST:
-                print(request.POST)
                 link = Link.objects.get(pk=request.POST['_delete'])
                 link.delete()
-                return redirect('links', page_slug=page.slug)
+                return redirect('links')
 
         return render(request, 'masters_app/links.html', context)
 
 
 class AlbumsView(DataMixin, View):
     def dispatch(self, request, *args, **kwargs):
-
         page = get_object_or_404(Page, slug=kwargs['page_slug'])
+        owner = False
+        if request.user.page == page:
+            owner = True
         albums = Album.objects.filter(page=page)
-        links = Link.objects.filter(page=page)
         title = f'{page}: Альбоми'
         page_obj = self.get_paginator(request, Album, 3, page=page)
         context = self.get_user_context(title=title,
                                         page=page,
                                         albums=albums,
-                                        links=links,
-                                        page_obj=page_obj)
+                                        page_obj=page_obj,
+                                        owner=owner)
         return render(request, 'masters_app/user_albums.html', context)
 
 
@@ -258,52 +275,53 @@ class AlbumAddView(DataMixin, CreateView):
     form_class = AlbumForm
     template_name = 'masters_app/album_constructor.html'
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.user = None
+
     def dispatch(self, request, *args, **kwargs):
-        self.page = Page.objects.get(slug=kwargs['page_slug'])
+        self.user = request.user
+        if not self.user.is_authenticated:
+            return redirect('login')
         return super(AlbumAddView, self).dispatch(request, args, kwargs)
 
     def get_context_data(self, *, object_list=None, **kwargs):
-        print(kwargs)
         context = super().get_context_data(**kwargs)
-        context.update(self.get_user_context(title="Створення альбому", mark='add'))
-        print(context)
+        context.update(self.get_user_context(title="Створення альбому", mark='add', page=self.user.page))
         return context
 
     def form_valid(self, form):
-        print('DATA', form.data)
-        print('CLEANED DATA PIC', form.cleaned_data['main_picture'])
+
         album_data = {
             'name': form.cleaned_data['name'],
             'description': form.cleaned_data['description'],
             'main_picture': form.cleaned_data['main_picture'],
-            'page': self.page
+            'page': self.user.page
         }
         Album.objects.create(**album_data)
-        return redirect('albums', self.page.slug)
+        return redirect('albums', self.user.page.slug)
 
 
 class AlbumEditView(DataMixin, View):
     def dispatch(self, request, *args, **kwargs):
 
-        page = get_object_or_404(Page, slug=kwargs['page_slug'])
-        form = None
-        album_id = kwargs['album_id']
-        title = 'Налаштування альбома'
+        album = get_object_or_404(Album, pk=kwargs['album_id'])
+        page = album.page
 
-        context = self.get_user_context(form=form, page=page, title=title, mark='edit')
+        # checking user
+        if request.user.page != page:
+            return redirect('home')
+
+        title = 'Налаштування альбома'
 
         # POST
         if request.method == 'POST':
-            album = get_object_or_404(Album, pk=album_id)
-
             if '_delete' in request.POST:
                 album.delete()
                 return redirect('albums', page_slug=page.slug)
 
             form = AlbumForm(instance=album, data=request.POST, files=request.FILES)
             if form.is_valid():
-                print('DATA:', form.data)
-                print('CLEANED DATA:', form.cleaned_data)
                 form.save()
                 return redirect('albums', page_slug=page.slug)
             else:
@@ -311,17 +329,83 @@ class AlbumEditView(DataMixin, View):
 
         # GET
         else:
-            album = get_object_or_404(Album, pk=kwargs['album_id'])
             form = AlbumForm(instance=album)
-            context = self.get_user_context(form=form,
-                                            page=page,
-                                            title=title,
-                                            mark='edit',
-                                            album_header=album.main_picture)
 
+        context = self.get_user_context(form=form,
+                                        page=page,
+                                        title=title,
+                                        mark='edit',
+                                        album_header=album.main_picture)
         return render(request, 'masters_app/album_constructor.html', context)
 
 
+class WorkAddView(DataMixin, CreateView):
+    form_class = WorkForm
+    template_name = 'masters_app/work_constructor.html'
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.album = None
+
+    def dispatch(self, request, *args, **kwargs):
+        self.album = get_object_or_404(Album, pk=kwargs['album_id'])
+
+        # checking user
+        if request.user.page != self.album.page:
+            return redirect('home')
+
+        return super(WorkAddView, self).dispatch(request, args, kwargs)
+
+    def get_context_data(self, *, object_list=None, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context.update(self.get_user_context(title="Додати роботу",
+                                             page=self.album.page,
+                                             album=self.album,
+                                             mark='add'))
+        return context
+
+    def form_valid(self, form):
+        work_data = {
+            'photo': form.cleaned_data['photo'],
+            'description': form.cleaned_data['description'],
+            'album': self.album,
+        }
+        Work.objects.create(**work_data)
+
+        return redirect('album', self.album.pk)
+
+
+class WorkEditView(DataMixin, View):
+    def dispatch(self, request, *args, **kwargs):
+
+        work = get_object_or_404(Work, pk=kwargs['work_id'])
+        album = work.album
+        page = album.page
+
+        # checking user
+        if request.user.page != page:
+            return redirect('home')
+
+        title = 'Змінити роботу'
+
+        # POST
+        if request.method == 'POST':
+            if '_delete' in request.POST:
+                work.delete()
+                return redirect('album', album_id=album.pk)
+            form = WorkForm(instance=work, data=request.POST, files=request.FILES)
+            if form.is_valid():
+                form.save()
+                return redirect('work', work_id=work.pk)
+            else:
+                form.add_error(None, 'ERROR')
+
+        # GET
+        else:
+            form = WorkForm(instance=work)
+
+        context = self.get_user_context(form=form, page=page, album=album, title=title, mark='edit')
+        return render(request, 'masters_app/work_constructor.html', context)
